#28: Cleanup
We no longer need the old code. To be clean, we create an extension of NSNotification.Name because Strings are flimsy. In order for the notification to be posted and observed, both strings must be the same. In our previous commit, we could've had a typo in one string and that would've broken the notification flow. This way, the string only exists in `Notifications+GuessMyBreed.swift`.